# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de eStreamPanel v2.0",
    "description": "eStreamPanel es un middleware de autorización para reproducción de transmisiones con Flussonic, además de gestionar planes, paquetes, suscriptores, entre muchas otras funciones.",
    "version": "2.0.0"
  },
  "host": "adminhyo.globaltv.lat",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token de autorización usando el esquema Bearer. Ejemplo: 'Authorization: Bearer {token}'"
    }
  },
  "paths": {
...
    "/auth/{user}/{password}": {
      "get": {
        "tags": [
          "Auth - Subscriber Info"
        ],
        "summary": "Get subscriber info",
        "description": "Obtiene la información básica del suscriptor.",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "password",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Consulta exitosa."
          },
          "403": {
            "description": "Error en la consulta."
          }
        }
      }
    },
    "/auth/{user}/{password}/categories": {
      "get": {
        "tags": [
          "Auth - Subscriber Categories"
        ],
        "summary": "Get subscriber categories",
        "description": "Obtiene la lista de categorias disponibles.",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "password",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Consulta exitosa."
          },
          "403": {
            "description": "Error en la consulta."
          }
        }
      }
    },
    "/auth/{user}/{password}/playlist/{responseFormat}/{videoFormat}": {
      "get": {
        "tags": [
          "Auth - Subscriber Playlist"
        ],
        "summary": "Get subscriber playlist",
        "description": "Obtiene la lista de reproducción del suscriptor.",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "password",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "responseFormat",
            "in": "path",
            "type": "string",
            "required": true,
            "enum": [
              "json",
              "m3u8"
            ],
            "description": "Formato de respuesta permitido: 'json' o 'm3u8'"
          },
          {
            "name": "videoFormat",
            "in": "path",
            "type": "string",
            "required": true,
            "enum": [
              "hls",
              "ts"
            ],
            "description": "Formato de video permitido: 'hls' o 'ts'"
          }
        ],
        "responses": {
          "200": {
            "description": "Consulta exitosa."
          },
          "403": {
            "description": "Error en la consulta."
          }
        }
      }
    },
    "/auth/token-validate": {
      "get": {
        "tags": [
          "Auth - Player Approval"
        ],
        "summary": "Validación de reproducción de video",
        "description": "Este endpoint debe ser configurado como mecanismo de autorización en Flussonic. Se encarga de validar los parámetros enviados por el reproductor antes de permitir el acceso al contenido, asegurando que el token y la sesión sean válidos.",
        "responses": {
          "200": {
            "description": "Reproducción autorizada."
          },
          "403": {
            "description": "Acceso denegado. La solicitud no cumplió con las validaciones."
          }
        }
      }
    },
...
  },
...
}